pyproject.toml
[project]
name = "cmsis-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
